refactor: hook 사용되는 곳으로 이동
