+++ src/views/ApplyPage/hooks/useGetDraft.tsx
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query';
 
 import { getDraft } from '../apis';
 
-import type { ApplyResponse } from '../../ApplyPage/types';
+import type { ApplyResponse } from '../types';
 import type { ErrorResponse } from '@type/errorResponse';
 import type { AxiosError, AxiosResponse } from 'axios';
