--- src/views/MyPage/apis.ts
@@ -1,7 +1,0 @@
-import tokenInstance from '@apis/tokenInstance';
-
-export const getMyInfo = async () => {
-  const res = await tokenInstance.get('/recruiting-auth/my');
-
-  return res;
-};
