+++ src/views/ResultPage/types.ts
@@ -1,20 +1,3 @@
-export interface MyRequest {
-  email: string;
-  season: number;
-  group: string;
-  password: string;
-}
-
-export interface MyResponse {
-  err: boolean;
-  season: number;
-  name: string;
-  part: string;
-  submit: boolean;
-  applicationPass: boolean;
-  finalPass: boolean;
-}
-
 export interface ScreeningResultResponse {
   err: boolean;
   season: number;
